Add send email capacity for trainee

+++ trainee/card.php
@@ -68,6 +68,9 @@ if ($reshook < 0)
 	setEventMessages($hookmanager->error, $hookmanager->errors, 'errors');
 
 
+//Select mail models is same action as presend
+if (GETPOST('modelselected')) $action = 'presend';
+
 /*
  * Actions delete
 */
@@ -361,17 +364,8 @@ if ($action == 'create_confirm' && ($user->rights->agefodd->creer || $user->righ
 /*
  * View
 */
-$title = ($action == 'nfcontact' || $action == 'create' ? $langs->trans("AgfMenuActStagiaireNew") : $langs->trans("AgfStagiaireDetail"));
-
-$form = new Form($db);
-$formcompany = new FormCompany($db);
-$formAgefodd = new FormAgefodd($db);
-
-/*
- * Action create
-*/
 if ($action == 'create' && ($user->rights->agefodd->creer || $user->rights->agefodd->modifier)) {
-    llxHeader('', $title);
+	llxHeader('', $title);
 	print "\n" . '<script type="text/javascript">
 		$(document).ready(function () {
 			$("input[type=radio][name=create_thirdparty]").change(function() {
@@ -488,7 +482,7 @@ if ($action == 'create' && ($user->rights->agefodd->creer || $user->rights->agef
 	print '</td>';
 	print '	</tr>';
 	print '<tr class="select_thirdparty_block"><td class="fieldrequired">' . $langs->trans("Company") . '</td><td colspan="3">';
-if (!empty($conf->global->AGEFODD_USE_SELECT_WITH_AJAX)) print $form->select_company(GETPOST('societe', 'int'), 'societe', '(s.client IN (1,3,2))', 'SelectThirdParty', 1);
+	if (!empty($conf->global->AGEFODD_USE_SELECT_WITH_AJAX)) print $form->select_company(GETPOST('societe', 'int'), 'societe', '(s.client IN (1,3,2))', 'SelectThirdParty', 1);
 	else print $form->select_thirdparty_list(GETPOST('societe', 'int'), 'societe', '(s.client IN (1,3,2))', 'SelectThirdParty', 1);
 	print '</td></tr>';
 
@@ -503,15 +497,15 @@ if (!empty($conf->global->AGEFODD_USE_SELECT_WITH_AJAX)) print $form->select_com
 	// Zip / Town
 	print '<tr class="create_thirdparty_block"><td>' . $langs->trans('Zip') . '</td><td>';
 	print $formcompany->select_ziptown($object->zip, 'zipcode', array (
-			'town',
-			'selectcountry_id',
-			'departement_id'
+		'town',
+		'selectcountry_id',
+		'departement_id'
 	), 6);
 	print '</td><td>' . $langs->trans('Town') . '</td><td>';
 	print $formcompany->select_ziptown($object->town, 'town', array (
-			'zipcode',
-			'selectcountry_id',
-			'departement_id'
+		'zipcode',
+		'selectcountry_id',
+		'departement_id'
 	));
 	print '</td></tr>';
 
@@ -639,7 +633,23 @@ if (!empty($conf->global->AGEFODD_USE_SELECT_WITH_AJAX)) print $form->select_com
 	print '</td></tr>';
 	print '</table>';
 	print '</form>';
-} else {
+}
+elseif (!empty($id) && $action == 'send')
+{
+	$object = new Agefodd_stagiaire($db);
+	$result = $object->fetch($id);
+	if($result>0){
+		// Actions to send emails
+		$actiontypecode = 'AC_OTH_AUTO';
+		$trigger_name = 'AGFTRAINEE_SENTBYMAIL';
+		$autocopy = 'MAIN_MAIL_AUTOCOPY_AGFTRAINEE_TO';
+		$trackid = 'agftrainee' . $object->id;
+		include DOL_DOCUMENT_ROOT . '/core/actions_sendmails.inc.php';
+	}
+
+}
+else
+{
 	// Affichage de la fiche "stagiaire"
 	if ($id) {
 		$agf = new Agefodd_stagiaire($db);
@@ -648,8 +658,8 @@ if (!empty($conf->global->AGEFODD_USE_SELECT_WITH_AJAX)) print $form->select_com
 		$head = trainee_prepare_head($agf);
 
 		if ($result > 0) {
-		    llxHeader('', $title);
-		    dol_fiche_head($head, 'card', $langs->trans("AgfStagiaireDetail"), 0, 'user');
+			llxHeader('', $title);
+			dol_fiche_head($head, 'card', $langs->trans("AgfStagiaireDetail"), 0, 'user');
 
 			// Affichage en mode "édition"
 			if ($action == 'edit') {
@@ -781,8 +791,8 @@ if (!empty($conf->global->AGEFODD_USE_SELECT_WITH_AJAX)) print $form->select_com
 			} else {
 				// Display in "view" mode
 
-			    dol_agefodd_banner_tab($agf, 'id');
-			    print '<div class="underbanner clearboth"></div>';
+				dol_agefodd_banner_tab($agf, 'id');
+				print '<div class="underbanner clearboth"></div>';
 
 				/*
 				* Confirmation de la suppression
@@ -836,10 +846,10 @@ if (!empty($conf->global->AGEFODD_USE_SELECT_WITH_AJAX)) print $form->select_com
 				print '<td>' . stripslashes($notes) . '</td></tr>';
 
 				if (! empty($conf->global->AGF_USE_REAL_HOURS)) {
-				    require_once ('../class/agefodd_session_stagiaire_heures.class.php');
-				    $agfssh = new Agefoddsessionstagiaireheures($db);
-				    print '<tr><td>' . $langs->trans('AgfTraineeHours') . '</td>';
-				    print '<td>' . $agfssh->heures_stagiaire_totales($agf->id) . ' h</td></tr>';
+					require_once ('../class/agefodd_session_stagiaire_heures.class.php');
+					$agfssh = new Agefoddsessionstagiaireheures($db);
+					print '<tr><td>' . $langs->trans('AgfTraineeHours') . '</td>';
+					print '<td>' . $agfssh->heures_stagiaire_totales($agf->id) . ' h</td></tr>';
 				}
 
 				if (! empty($extrafields->attribute_label)) {
@@ -858,18 +868,29 @@ if (!empty($conf->global->AGEFODD_USE_SELECT_WITH_AJAX)) print $form->select_com
 
 			print '<div class="tabsAction">';
 
-			if ($action != 'create' && $action != 'edit' && $action != 'nfcontact') {
-			    if ($user->rights->agefodd->creer || $user->rights->agefodd->modifier) {
-			        print '<a class="butAction" href="' . $_SERVER['PHP_SELF'] . '?action=edit&id=' . $id . '">' . $langs->trans('Modify') . '</a>';
-			    } else {
-			        print '<a class="butActionRefused" href="#" title="' . dol_escape_htmltag($langs->trans("NotAllowed")) . '">' . $langs->trans('Modify') . '</a>';
-			    }
-			    if ($user->rights->agefodd->creer || $user->rights->agefodd->modifier) {
-			        print '<a class="butActionDelete" href="' . $_SERVER['PHP_SELF'] . '?action=delete&id=' . $id . '">' . $langs->trans('Delete') . '</a>';
-			    } else {
-			        print '<a class="butActionRefused" href="#" title="' . dol_escape_htmltag($langs->trans("NotAllowed")) . '">' . $langs->trans('Delete') . '</a>';
-			    }
+			if ($action != 'create' && $action != 'edit' && $action != 'nfcontact' && $action != 'presend') {
+
+				// Send
+				if ($user->rights->agefodd->creer || $user->rights->agefodd->modifier) {
+					print '<div class="inline-block divButAction"><a class="butAction" href="' . $_SERVER["PHP_SELF"] . '?id=' . $agf->id . '&action=presend&mode=init#formmailbeforetitle">' . $langs->trans('SendMail') . '</a></div>';
+				} else {
+					print '<div class="inline-block divButAction"><a class="butActionRefused" href="#">' . $langs->trans('SendMail') . '</a></div>';
+				}
+
+				if ($user->rights->agefodd->creer || $user->rights->agefodd->modifier) {
+					print '<a class="butAction" href="' . $_SERVER['PHP_SELF'] . '?action=edit&id=' . $id . '">' . $langs->trans('Modify') . '</a>';
+				} else {
+					print '<a class="butActionRefused" href="#" title="' . dol_escape_htmltag($langs->trans("NotAllowed")) . '">' . $langs->trans('Modify') . '</a>';
+				}
+
+				if ($user->rights->agefodd->creer || $user->rights->agefodd->modifier) {
+					print '<a class="butActionDelete" href="' . $_SERVER['PHP_SELF'] . '?action=delete&id=' . $id . '">' . $langs->trans('Delete') . '</a>';
+				} else {
+					print '<a class="butActionRefused" href="#" title="' . dol_escape_htmltag($langs->trans("NotAllowed")) . '">' . $langs->trans('Delete') . '</a>';
+				}
 			}
+
+
 
 		} else {
 			setEventMessage($agf->error, 'errors');
@@ -878,12 +899,33 @@ if (!empty($conf->global->AGEFODD_USE_SELECT_WITH_AJAX)) print $form->select_com
 		}
 	}
 }
+$title = ($action == 'nfcontact' || $action == 'create' ? $langs->trans("AgfMenuActStagiaireNew") : $langs->trans("AgfStagiaireDetail"));
+
+$form = new Form($db);
+$formcompany = new FormCompany($db);
+$formAgefodd = new FormAgefodd($db);
+
+/*
+ * Action create
+*/
 
 $parameters = array();
 $reshook = $hookmanager->executeHooks('formObjectOptions', $parameters, $agf, $action); // Note that $action and $object may have been modified by hook
 print $hookmanager->resPrint;
 
 print '</div>';
+
+
+if ($id) {
+
+	// Presend form
+	$modelmail = 'agf_trainee';
+	$defaulttopic = 'AgfSendEmailTrainee';
+	$diroutput = $conf->agefodd->multidir_output[$agf->entity];
+	$trackid = 'agftrainee' . $agf->id;
+
+	include __DIR__ . '/card_presend.tpl.php';
+}
 
 llxFooter();
 $db->close();
